+++ crates/cxx-qt-lib/src/gui/init.cpp
@@ -8,7 +8,15 @@
 #include <cxx-qt-lib/qlist.h>
 #include <cxx-qt-lib/qvector.h>
 
-static const int register_QList_QColor =
-  qRegisterMetaType<::QList_QColor>("QList_QColor");
-static const int register_QVector_QColor =
-  qRegisterMetaType<::QVector_QColor>("QVector_QColor");
+#include <mutex>
+
+extern "C" bool
+init_cxx_qt_lib_gui()
+{
+  static std::once_flag flag;
+  std::call_once(flag, []() {
+    qRegisterMetaType<::QList_QColor>("QList_QColor");
+    qRegisterMetaType<::QVector_QColor>("QVector_QColor");
+  });
+  return true;
+}